==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/01_data_profiling_complete.Rmd <==
---
title: "01_data_profiling_complete"
author: "Crystal"
output: 
  rmarkdown::html_document:
    toc: true
    toc_float: true
    highlight: tango
    theme: journal
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, include=FALSE, paged.print=FALSE}
# load packages 
for (pkg in c("tidyverse", "data.table", "stringr", "stringi", "mosaic", "dplyr", "readr","gt")) {
  library(pkg, character.only = TRUE)
}

get_db_conn <-
  function(db_name = "sdad",
           db_host = "postgis1",
           db_port = "5432",
           db_user = Sys.getenv("db_usr"),
           db_pass = Sys.getenv("db_pwd")) {
    RPostgreSQL::dbConnect(
      drv = RPostgreSQL::PostgreSQL(),
      dbname = db_name,
      host = db_host,
      port = db_port,
      user = db_user,
      password = db_pass
    )
  }

con <- get_db_conn()


pers <- DBI::dbGetQuery(con, "
                        SELECT * FROM bgt_res.pers
    "
)

DBI::dbDisconnect(con)

# us zipcode data from here: https://simplemaps.com/data/us-zips
uszips <- read_csv("~/git/dspg20career/src/burningglass/uszips.csv")%>%
  select(zip, city, state_id, state_name, county_name)

# zipcodes in 47900 msa code
# ZCTA (zip code tabulated area) to MSA relation file
zips47900 <- as.vector(read.delim("~/git/dspg20career/src/geography/zips47900.txt")$ZCTA5)
```

# Objective
Provide data profiling on the complete burning glass dataset. 

# Summary
In pers table, `r sum(is.na(pers$msa))/nrow(pers)`% (N= `r sum(is.na(pers$msa))`) missing in the msa variable, which is an identifier for geographic location. `r length(grep("47900", pers$msa, fixed = T))` users came from Washington-Arlington-Alexandria, DC-VA-MD-WV Metropolitan Statistical Area, identified by code: 47900.    

Among 687129 people, only 0.30 % (N=2093) of people had duplicated entries in msa. Nearly 100% (N=687127) have zipcode. We have 96.65% (N=664116) matched the zipcodes we approximated that match 47900 msa code.

To summarize, "msa" is a reliable geographic variable for us to identify people from the DC metropolitan area. We identify anyone who entered 47900 in the msa code as the population who live in the DC metropolitan area.

```{r message=FALSE, warning=FALSE, paged.print=FALSE}
# function for data profiling
data_profiling <- function(df){
summary_table <- tibble(var = names(df),
       variable_type = map_chr(.x = df, .f = function(col) class(x = col)),                 
       num_unique = map_int(.x = df, .f = function(col) length(x = unique(x = col))),
       num_missing = map_int(.x = df, .f = function(col) sum(x = is.na(x = col)))) %>%
  mutate(perc_missing = round(x = 100 * num_missing / nrow(x = .), digits = 2L))

return(summary_table)
}
```

```{r message=FALSE, warning=FALSE}
pers%>%
  data_profiling()%>%
  gt()
DC_pers <- pers%>%
  filter(grepl("47900", msa) == T)
```

# Identify DC Metropolitan Area

## Identify duplicated MSA code
```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
matches <- regmatches(DC_pers$msa, gregexpr("[[:digit:]]+", DC_pers$msa))
msa_list <- unique(unlist(matches))
msa_DC <- c("47900", "548")
msa_not_DC <- setdiff(msa_list, msa_DC)

dup_detect <- c()
for (i in (1:length(matches))){
  item <- matches[[i]]
  if(sum(item %in% msa_not_DC == T) > 2){
    dup_detect[i] = T
  }else{
    dup_detect[i] = F
  }
}

nrow(DC_pers)

table(dup_detect) #2093 ppl had msa duplicates

DC_pers_no_dup <- cbind(DC_pers, dup_detect)

DC_pers_no_dup <- DC_pers%>%
  filter(!dup_detect)
```

```{r eval=FALSE, include=FALSE}
#library(maditr)
#msa <- fread(input = "https://www2.census.gov/programs-surveys/cbp/technical-documentation/reference/metro-area-geography-reference/msa_county_reference17.txt")
# https://www2.census.gov/programs-surveys/cbp/technical-documentation/reference/state-county-geography-reference/georef17.txt for States + Counties (FIPS)
```

Among people who only reported "47900" in their msa, there are 5033 unique self-reported zipcodes. Then, I matched the zipcode with zipcodes on US Zip Codes Database (using authoritative sources including the U.S. Postal Service™, U.S. Census Bureau, National Weather Service, American Community Survey, and the IRS). 28.00% of these 5033 zipcodes (N=1409) don't match with the zipcode database. 
```{r message=FALSE, warning=FALSE, paged.print=FALSE}
DC_pers <- DC_pers%>%
  mutate(zipcode_clean = substr(zipcode,1,5))
table(is.na(DC_pers$zipcode_clean))

DC_pers <- filter(DC_pers, !is.na(DC_pers$zipcode_clean))

# DCzipcode <- DC_pers%>%
#   group_by(zipcode_clean)%>%
#   summarize(N=n())%>%
#   left_join(uszips, by = c("zipcode_clean" = "zip"))%>%
#   arrange(desc(N))

# DCzipcode%>%
#   group_by(state_id)%>%
#   summarize(number_zips=n())%>%
#   arrange(desc(number_zips))%>%
#   gt()%>%
#    tab_header(
#     title = "states zipcode in 47900 msa code"
#   )
zips47900 <- as.character(zips47900)

ttl_pers <-  sum(DC_pers$N)

DC_pers%>%
  mutate(match_47900zip = if_else(DC_pers$zipcode_clean %in% zips47900, T, F))%>%
  group_by(match_47900zip)%>%
  summarize(sum_match = n())%>%
  mutate(percent_match = sum_match/nrow(DC_pers)*100)%>%
  gt()
```




# Subset Burning Glass data using DC/VA/MD zipcodes
Alternatively, we could us zipcodes identified in the zipcode database. There are 52 zipcodes identified in the District of Columbia. We could add more states if we want. For example 893 zipcodes from VA, 468 zipcodes from Maryland. In total, there are 1416 zipcodes in DC/VA/MD. In Burning Glass data, 1,307,546 people are located in DC/VA/MD, 84,337 people located in DC using zipcode identifier.
```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
table(uszips$state_id)
# only DC
dc_zips <- as.vector(filter(uszips, state_id=="DC")$zip)

dc_zips <- uszips%>%
  getElement(name = c("zip", "state_id"))%>%
  subset("state_id" == "DC")

pers_dc <- pers%>%
  filter(zipcode %in% dc_zips)
nrow(pers_dc)

# DC, VA, MD
dc_va_md_zips <- uszips%>%
  filter(state_id %in% c("DC", "VA", "MD"))

dc_va_md_zips <- as.vector(dc_va_md_zips$zip)
length(dc_va_md_zips)

pers_dc_va_md <- pers%>%
  filter(zipcode %in% dc_va_md_zips)
nrow(pers_dc_va_md)
```
