==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/02_data_profiling_DC.Rmd <==
---
title: "Data Profiling DC"
output: html_document
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
#knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, warning=FALSE, include=FALSE, paged.print=FALSE}
# load packages 
for (pkg in c("tidyverse", "data.table", "stringr", "stringi", "mosaic", "dplyr", "readr","gt", 'lubridate')) {
  library(pkg, character.only = TRUE)
}

get_db_conn <-
  function(db_name = "sdad",
           db_host = "postgis1",
           db_port = "5432",
           db_user = Sys.getenv("db_usr"),
           db_pass = Sys.getenv("db_pwd")) {
    RPostgreSQL::dbConnect(
      drv = RPostgreSQL::PostgreSQL(),
      dbname = db_name,
      host = db_host,
      port = db_port,
      user = db_user,
      password = db_pass
    )
  }

con <- get_db_conn()

cert <- DBI::dbGetQuery(con, "SELECT *
FROM bgt_res.cert A
JOIN bgt_res.id_msa_dc B
ON A.id = B.id"
)

ed <- DBI::dbGetQuery(con, "SELECT *
FROM bgt_res.ed A
JOIN bgt_res.id_msa_dc B
ON A.id = B.id"
)

job <- DBI::dbGetQuery(con, "SELECT *
FROM bgt_res.job A
JOIN bgt_res.id_msa_dc B
ON A.id = B.id"
)
  
pers <- DBI::dbGetQuery(con, "SELECT *
FROM bgt_res.pers A
JOIN bgt_res.id_msa_dc B
ON A.id = B.id"
)

skill <- DBI::dbGetQuery(con, "SELECT *
FROM bgt_res.skill A
JOIN bgt_res.id_msa_dc B
ON A.id = B.id"
)

DBI::dbDisconnect(con)

```


```{r}
# function for data profiling
data_profiling <- function(df){
summary_table <- tibble(var = names(df),
       variable_type = map_chr(.x = df, .f = function(col) class(x = col)),                 
       num_unique = map_int(.x = df, .f = function(col) length(x = unique(x = col))),
       num_missing = map_int(.x = df, .f = function(col) sum(x = is.na(x = col)))) %>%
  mutate(perc_missing = round(x = 100 * num_missing / nrow(x = .), digits = 2L))
return(summary_table)
}
```


# Table: Cert
In total, there are `r nrow(cert)` entries in cert table. There are two types of certificates: certification and license. On average, each person received 2 certificates (sd = 1.68). 
```{r message=FALSE, warning=FALSE, paged.print=FALSE}
#check that id columns are identical and drop the final one
ncol(cert)
identical(cert[[1]],cert[[5]])
cert <- cert[-5]

data_profiling(cert)

cert <- cert%>%
  mutate(type = if_else(type == "license", "License", type))

table(cert$type)

cert_n <- cert%>%
  group_by(id)%>%
  summarize(n_cert = n())

ggplot(cert_n, aes(x = n_cert)) +
  geom_histogram(binwidth = 1.3) +
  labs(title = "Distribution of the number of certificates people received", x = "number of certificates")
favstats(cert_n$n_cert)

# top 10 certificates
cert_name <- cert%>%
  group_by(name)%>%
  summarize(n_cert_name = n())%>%
  arrange(desc(n_cert_name))%>%
  top_n(10)%>%
  rename("number of people who received the certificate" = "n_cert_name")
```



# Table: Ed

```{r message=FALSE, warning=FALSE, paged.print=FALSE}
#check that id columns are identical and drop one
ncol(ed)
identical(ed[[1]],ed[[13]])
ed <- ed[-13]

data_profiling(ed)
```


# Table: Pers
```{r message = FALSE, warning = FALSE}
#check that id columns are identical and drop one
ncol(pers)
identical(pers[[1]],pers[[10]])
pers <- pers[-10]

data_profiling(pers)

# Preliminary profiling for pers table. 

pers %>% group_by(msa) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))

# Checking on MSA. This column is messy and it looks like multiple values are being captured. 

pers %>% group_by(gender) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))

# Checking on gender. There are slightly more male than female resumes, but overall pretty even distribution.

pers %>% group_by(zipcode) %>% summarise(n = n()) %>% mutate(freq = n / sum(n)) %>% arrange(-freq) %>% head(15)

# Checking on zip code. The distribution is pretty even. The top 15 are all in the DC metropolitan area. 

pers %>% group_by(noofjobs) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))
ggplot(pers, aes(x = noofjobs)) +
  geom_histogram(bins = 19)

# Checking on the number of jobs. 
```



# Table: Job
```{r message = FALSE, warning = FALSE}
#check that id columns are identical and drop one
ncol(job)
identical(job[[1]],job[[10]])
job <- job[-10]

data_profiling(job)

# Preliminary profiling for job table. 

job %>% count(onet) %>% arrange(-n) %>% head(15)

# Checking on onet. There may be some issues with this variable since the number of unique values (1046) is more than the number of unique onet codes online (1016)

year_start <- job %>% mutate(year = year(startdate)) %>% filter(!is.na(year)) %>% group_by(year) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))

# Checking on startdate. Some a few observations seem impossible (1900) and some might be unlikely. 

ggplot(year_start, aes(x = year, y = n)) +
  geom_smooth() +
  xlim(2000, 2020)

# The distribution of startdates from 2000-2020 looks as expected. 
       
year_end <- job %>% mutate(year = year(enddate)) %>% filter(!is.na(year)) %>% group_by(year) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))

ggplot(year_end, aes(x = year, y = n)) +
         geom_smooth() +
         xlim(2000, 2020)

# The distribution of enddates is concentrated after 2015. We might expect this since the data is resumes collected from jobseekers around this time.

job %>% mutate(month = month(startdate)) %>% filter(!is.na(month)) %>% group_by(month) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))

job %>% mutate(day = day(startdate)) %>% filter(!is.na(day)) %>% group_by(day) %>% summarise(n = n()) %>% mutate(freq = n / sum(n))

# For month there is a greater frequency of startdates in January. For day the frequency is greatest for the first of the month. 

table(ed$degreetype)

degree_clean <- ed%>%
  filter(!is.na(degreetype))%>%
  mutate(degree_somehs = if_else(str_detect(string = degreetype, 
                              pattern = "\\b(?i)(10|11|9)\\b"), T,F))%>%
  mutate(degree_hs = if_else(str_detect(string = degreetype, 
                              pattern = "\\b(?i)(12|High School|ged)\\b"), T,F))%>%
  mutate(degree_bachelor = if_else(str_detect(string = degreetype, 
                              pattern = "\\b(?i)(Bachelor|bachelor|Bachelors|BS|bs|AA|A.A|Undergraduate|undergraduate|postgraduate|Associate|associate)\\b"), T,F))%>%
  mutate(degree_master = if_else(str_detect(string = degreetype, 
                              pattern = "\\b(?i)(master|Master|MBA|M.S|MS|MD)\\b"), T,F))%>%
  mutate(degree_doctor = if_else(str_detect(string = degreetype, 
                              pattern = "\\b(?i)(phd|Ph.D|postdoc)\\b"), T,F))%>%
  mutate(degree_highest = if_else(degree_doctor == T, "doctor", 
                                  if_else(degree_master == T, "master",
                                          if_else(degree_bachelor == T, "bachelor", if_else(degree_hs == T, "hs", 
                                                                                            if_else(degree_somehs == T, "somehs", "others"))))))



table(degree_clean$degree_highest)
```

In total, there are `r nrow(ed)` entries in ed table. After a preliminary clean up of degree, we identified `r round((nrow(degree_clean)-267101)/(nrow(degree_clean))* 100, digits = 2) `% of self reported degree type into 5 categories: some high school, high school, bachelor, master, and doctor's degree. 

# Table: Skill

```{r}
#check that id columns are identical and drop one
ncol(skill)
identical(skill[[1]],skill[[9]])
skill <- skill[-9]

data_profiling(skill)
```

In total, there are `r nrow(skill)` entries in cert table, representing `r length(unique(skill$id))` unique IDs and `r length(unique(skill$skillid))` skills. The number of unique enumerated skills (written out) is `r length(unique(skill$skill))` which means that the number of skillids does not match the number of skills. It is possible that they were coded to have some overlap, though it's strange that the numbers are so close and yet not equal.

Below is a list of the top 10 skills:

```{r}
top_10 <- skill %>%
  group_by(skill) %>%
  summarize(n = n()) %>%
  arrange(desc(n)) %>%
  head(10)
top_10
```

Below is a list of the bottom 15 skills. These are much messier and there are many more skills with only one entry.

```{r}
bottom_15 <- skill %>%
  group_by(skill) %>%
  summarize(n = n()) %>%
  arrange((n)) %>%
  head(15)
bottom_15
```

Below is a plot of the `isbaseline` variable:

```{r}
skill %>%
  group_by(isbaseline) %>%
  summarize(n=n()) %>%
  ggplot() +
  geom_bar(aes(x = isbaseline, y = n), stat = 'identity') +
  geom_text(aes(x = isbaseline, y = n, label = n), vjust = -0.5)
```

Below is a plot of the `issoftware` variable:
```{r}
skill %>%
  group_by(issoftware) %>%
  summarize(n=n()) %>%
  ggplot() +
  geom_bar(aes(x = issoftware, y = n), stat = 'identity') +
  geom_text(aes(x = issoftware, y = n, label = n), vjust = -0.5)
```

Below is a plot of the `isspecialized` variable:
```{r}
skill %>%
  group_by(isspecialized) %>%
  summarize(n=n()) %>%
  ggplot() +
  geom_bar(aes(x = isspecialized, y = n), stat = 'identity') +
  geom_text(aes(x = isspecialized, y = n, label = n), vjust = -0.5)
```

```{r}
uva_vet_jobs %>% filter(id == 5930181) %>% select(-rowname, -id.1)

ed %>% rownames_to_column() %>% filter(id == 5930181) %>% select(-rowname, -id.1)

skill %>% rownames_to_column() %>% filter(id == 5930181) %>% select(-rowname, -id.1)

cert %>% rownames_to_column() %>% filter(id == 5930181) %>% select(-rowname, -id.1)

pers %>% rownames_to_column() %>% filter(id == 5930181) %>% select(-rowname, -id.1)

sample.data <- job %>% rownames_to_column() %>% filter(id == 5930181)
```
