==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/05_acs_brfss_bgt_degree_comparison.Rmd <==
---
title: "05_comparison_acs_brfss_bgt"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
for (pkg in c("tidyverse",  "data.table", "R.utils", "maditr", "stringr", "stringi", "dplyr", "ggplot2", "lubridate", "gt", "DataExplorer", "tidycensus")) {
  library(pkg, character.only = TRUE)
}

#double check your working directory
source("~/git/DSPG2020/career/src/burningglass/03_clean_bgt.R", echo = T, prompt.echo = "", spaced = F)
bg_clean_ls <- clean_bg(job = job, ed = ed, pers = pers, onet = onet)

bg_all_demographic <- bg_clean_ls$bg_all_demographic
bg_vet_demographic <- bg_clean_ls$bg_vet_demographic
bg_all_job <- bg_clean_ls$bg_all_job
bg_vet_job <- bg_clean_ls$bg_vet_job
```


# ACS
```{r}
acs_vet_ed <- read_csv("~/git/DSPG2020/career/src/acs/vet_ed.csv")


colnames(acs_vet_ed) <- c("education", "count", "perc", "veteran")

sample_size <- sum(acs_vet_ed$count)

acs_vet_ed <- acs_vet_ed%>%
  mutate(education = if_else(education == "less_hs", "less than high school", 
                          if_else(education== "hs", "high school", 
                                  if_else(education == "some_college", "some college/associate", "bachelor and more"))))%>%
  select(-count)%>%
  mutate(perc = perc * 100)%>%
  mutate(data = "American Community Survey \n (ACS), 2018 Five-year estimates \n D.C. MSA, N = 4,144,490")
acs_vet_ed
```

#BRFSS
less than high school, high school, some college, bachelor and more
```{r}
brfss2018 <- read_csv("~/git/DSPG2020/career/src/brfss/brfss2018.csv")%>%
  select(-X1)

table(brfss2018$education)
brfss2018_vet_edu <- brfss2018 %>%
  filter(education != "Refused")%>%
  mutate(education = if_else(education == "College 1 year to 3 years (Some college or technical school)", "some college/associate",
                             if_else(education == "College 4 years or more (College graduate)", "bachelor and more",
                                     if_else(education == "Grade 12 or GED (High school graduate)", "high school",
                                           "less than high school"))))%>%
  mutate(veteran = if_else(veteran == "Yes", "Veteran",
                           if_else(veteran == "No", "Non-Veteran", "others")))

brfss_vet_edu <- table(brfss2018_vet_edu$veteran, brfss2018_vet_edu$education)%>%
  as.data.frame()
sample_size <- sum(brfss_vet_edu$Freq)

brfss_vet_edu <- brfss_vet_edu%>%
  rename("veteran" = "Var1", "education" = "Var2")%>%
  filter(veteran != "others")%>%
  group_by(veteran)%>%
  summarize(education=education, perc = Freq / sum(Freq) * 100)%>%
  filter(!is.na(education))%>%
  mutate(data = "Behavioral Risk Factor Surveillance System \n (BRFSS), 2018 \n National , N = 435,300")
brfss_vet_edu
```

#BGT
```{r}
bg_all_edu <- bg_all_demographic%>%
  rename(education = degree_highest)%>%
  filter(education != "other")%>%
  mutate(education = if_else(education %in% c("certificate", "somehs"), "less than high school",
                             if_else(education == "highschool", "high school",
                                     if_else(education == "associate", "some college/associate", "bachelor and more"))))%>%
  mutate(veteran = if_else(veteran == "veteran", "Veteran", "Non-Veteran"))

vet_edu <- table(bg_all_edu$veteran, bg_all_edu$education)%>%
  as.data.frame()

sample_size <- sum(vet_edu$Freq)

vet_edu <- vet_edu %>%
  rename("veteran" = "Var1", "education" = "Var2")%>%
  group_by(veteran)%>%
  summarize(education=education, perc = Freq / sum(Freq) * 100)%>%
  filter(!is.na(education))%>% 
  mutate(data = "Burning Glass \n 2016-2018 \n D.C. MSA, N = 620,497")
vet_edu
```


#Plot all
less than high school, high school, some college/associate, bachelor and more

```{r, fig.width = 15}
library(ggthemes)

uva_color_palette <- 
c("#232D4B", #space cadet
  "#2C4F6B", #indigo dye
  "#0E879C", #blue munsell
  "#60999A", #cadet blue
  "#D1E0BF", #tea green
  "#D9E12B", #pear
  "#E6CE3A", #citrine
  "#E6A01D", #marigold
  "#E57200" #princeton orange
)


vet_edu_all <- rbind(acs_vet_ed,brfss_vet_edu, vet_edu)
vet_edu_all

vet_edu_all <- vet_edu_all%>%
  mutate(education = if_else(education == "less than high school", "less than \n high school", if_else(education == "some college/associate", "some college \n or associate", if_else(education == "bachelor and more", "bachelor \n and more", education))))


vet_edu_all$education <- factor(vet_edu_all$education , levels = c("less than \n high school","high school","some college \n or associate",  "bachelor \n and more"))


ggplot(vet_edu_all, aes(x = education, y = perc, fill = data))+
  geom_bar(stat = "identity",position = 'dodge')+
  scale_fill_brewer(palette = "Set2")+
  facet_wrap(~veteran) +
  labs(title = "Education Comparison in ACS, BRFSS, BGT", y= "percent")+ 
  scale_fill_manual(values = c(uva_color_palette[1],uva_color_palette[4],uva_color_palette[8])) + 
  theme_hc()+
  theme(legend.position="top") 
#save size: 980x500
```
