==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/07_ward_clustering.rmd <==
---
title: "Exploring Ward Clustering"
output: html_document
author: Joanna Schroeder
---

## What is Ward's Method?
Ward’s method (a.k.a. Minimum variance method or Ward’s Minimum Variance Clustering Method) is an alternative to single-link clustering. Popular in fields like linguistics, it’s liked because it usually creates compact, even-sized clusters. Like most other clustering methods, Ward’s method is computationally intensive. However, Ward’s has significantly fewer computations than other methods. The drawback is this usually results in less than optimal clusters. That said, the resulting clusters are usually good enough for most purposes.

Like other clustering methods, Ward’s method starts with n clusters, each containing a single object. These n clusters are combined to make one cluster containing all objects. At each step, the process makes a new cluster that minimizes variance, measured by an index called E (also called the sum of squares index).

## What is Partitioning Around Medoids (PAM)?
K-medoids or partitioning around medoids (PAM) algorithm is a clustering algorithm reminiscent of the k-means algorithm. Both the k-means and k-medoids algorithms are partitional (breaking the dataset up into groups) and both attempt to minimize the distance between points labeled to be in a cluster and a point designated as the center of that cluster.

```{r include = TRUE, echo = FALSE, warning = FALSE, message = FALSE}
library(tidyverse)
library(class)
library(TraMineR)
library(dplyr)
library(cluster)

## Load data and sample

set.seed(01301998)

sts_vet <- read.csv("~/git/DSPG2020/career/data/sts_vet.csv", row.names = 1)

sts_vet_LDEL <- seqdef(sts_vet, cpal = c("#D1E0BF",  "#60999A", "#0E879C","#2C4F6B", "#232D4B", "gray"), left="DEL", gaps="NA", right="DEL")
sts_vet_LNA <- seqdef(sts_vet, cpal = c("#D1E0BF",  "#60999A", "#0E879C","#2C4F6B", "#232D4B", "gray", "black"), left="NA", gaps="NA", right="DEL")
```

## Sequence descriptives
```{r include = TRUE, echo = FALSE, warning = FALSE, message = FALSE}
# Seqmtplot displays the mean time spent in each state
#seqmtplot(sts_vet, main = "Mean time distribution plot", ylim = c(0, 10), border = NA, with.legend= FALSE)

# Seqdplot represents the sequence of the cross-sectional state frequencies by position
#seqdplot(sts_vet_LDEL, main = "Sequence distribution plot",  xlim = c(0, 46), border = NA, with.legend = "right")
#seqdplot(sts_vet_LNA, main = "Sequence distribution plot",  xlim = c(0, 46), border = NA, with.legend = "right")

# Seqfplot displays the most frequent sequences
#seqfplot(sts_vet, main = "Sequence frequency plot", with.legend = TRUE ,pbarw = TRUE)

# Transition matrix
tr_LDEL <- seqtrate(sts_vet_LDEL)
tr_LNA <- seqtrate(sts_vet_LNA)
round(tr_LDEL, 2)
round(tr_LNA, 2)
```


## Clustering and dendrogram
```{r include = TRUE, echo = FALSE, warning = FALSE, message = FALSE}
# Creating distance matrix using optimal matching, including missing (unemployment) values
dist_om_LDEL <- seqdist(sts_vet_LDEL, method = "OM", indel = 1, sm = "TRATE", with.missing = TRUE)
dis_om_LNA <- seqdist(sts_vet_LNA, method = "OM", indel = 1, sm = "TRATE", with.missing = TRUE)

# Implementing clustering, using Ward's method
#clusterward1 <- agnes(dist.om1, diss = TRUE, method = "ward")

# Plot the dendrogram
#plot(clusterward1, which.plot = 2)

# The first wards cluster is the left delete option
clusterward_LDEL <- readRDS("~/git/DSPG2020/career/data/clusterward_02.rds")
plot(clusterward_LDEL, which.plot = 2)

# The second wards cluster is the left NA option
clusterward_LNA <- readRDS("~/git/DSPG2020/career/data/clusterward_leftNA.rds")
plot(clusterward_LNA, which.plot = 2)
```


### LDEL Clusters
```{r include = TRUE, echo = TRUE, warning = FALSE, message = FALSE, fig.height=}
plot_clusters <- function(data, n) {
  cl1.n <- cutree(data, k = n)
  cl1.nfac <- factor(cl1.n, labels = paste("Type", 1:n))
  table(cl1.n)
  seqrplot(sts_vet_LDEL, diss = dist_om_LDEL, group = cl1.nfac, border = NA)
  seqdplot(sts_vet_LDEL, group = cl1.nfac, border = NA)
  seqfplot(sts_vet_LDEL, group = cl1.nfac, border = NA)
  seqmtplot(sts_vet_LDEL, group = cl1.nfac, border = NA)
  seqHtplot(sts_vet_LDEL, group = cl1.nfac, border = NA)
}
par(mar = c(2,2,2,2))
plot_clusters(clusterward_LDEL, 3)
plot_clusters(clusterward_LDEL, 4)
plot_clusters(clusterward_LDEL, 5)
plot_clusters(clusterward_LDEL, 6)
plot_clusters(clusterward_LDEL, 7)
plot_clusters(clusterward_LDEL, 8)

# Number of sequences in each cluster

# Seqrplot displays a reduced, non redundant set of representative sequences extracted from the provided state sequence object and sorted according to a representativeness criterion

# Seqdplot represents the sequence of the cross-sectional state frequencies by position (time point)

# Seqfplot displays the most frequent sequences, each one with an horizontal stack bar of its successive states

# Seqmtplot displays the mean time spent in each state

# Seqhtplot displays the evolution over positions of the cross-sectional entropies ( entropy is 0 when all cases are in the same state and is maximal when the same proportion of cases are in each state; the entropy can be seen as a measure of the diversity of states observed at the considered time point)

```


### LNA Clusters
```{r include = TRUE, echo = TRUE, warning = FALSE, message = FALSE, fig.height=}
plot_clusters <- function(data, n) {
  cl1.n <- cutree(data, k = n)
  cl1.nfac <- factor(cl1.n, labels = paste("Type", 1:n))
  table(cl1.n)
  seqrplot(sts_vet_LNA, diss = dis_om_LNA, group = cl1.nfac, border = NA)
  seqdplot(sts_vet_LNA, group = cl1.nfac, border = NA)
  seqfplot(sts_vet_LNA, group = cl1.nfac, border = NA)
  seqmtplot(sts_vet_LNA, group = cl1.nfac, border = NA)
  seqHtplot(sts_vet_LNA, group = cl1.nfac, border = NA)
}
par(mar = c(2,2,2,2))
plot_clusters(clusterward_LNA, 3)
plot_clusters(clusterward_LNA, 4)
plot_clusters(clusterward_LNA, 5)
plot_clusters(clusterward_LNA, 6)
plot_clusters(clusterward_LNA, 7)
plot_clusters(clusterward_LNA, 8)

```



### First 10 years
```{r include = TRUE, echo = TRUE, warning = FALSE, fig.height=10, message = FALSE}
library(naniar)

# Redefining sequence to delete left and gaps and subsetting to first 10 years. So this gives us first 10 years of employment
sts_vet_TEN <- seqdef(sts_vet, cpal = c("#D1E0BF",  "#60999A", "#0E879C","#2C4F6B", "#232D4B"), left="DEL", gaps="DEL", right="DEL")
sts_vet_TEN <- sts_vet_TEN[,1:10]
sts_vet_TEN <- sts_vet_TEN %>% replace_with_na_all(condition = ~.x == "%")
sts_vet_TEN <- sts_vet_TEN[complete.cases(sts_vet_TEN[1:10]),]

sts_vet_TEN <- seqdef(sts_vet_TEN, cpal = c("#D1E0BF",  "#60999A", "#0E879C","#2C4F6B", "#232D4B"), left="DEL", gaps="DEL", right="DEL")

# 4998 cases
dis_om_TEN <- seqdist(sts_vet_TEN, method = "OM", indel = 1, sm = "TRATE", with.missing = TRUE)

# Implementing clustering, using Ward's method
#clusterward_TEN <- agnes(dis_om_TEN, diss = TRUE, method = "ward")
clusterward_TEN <- readRDS(file = "~/git/DSPG2020/career/data/clusterward_TEN.rds")
#saveRDS(clusterward_TEN, file = "~/git/DSPG2020/career/data/clusterward_TEN.rds")
plot(clusterward_TEN, which.plot = 2)

seqmtplot(sts_vet_TEN, main = "Mean time distribution plot", ylim = c(0, 10), border = NA, with.legend= FALSE)
seqdplot(sts_vet_TEN, main = "Sequence distribution plot", border = NA, with.legend = "right")
seqdplot(sts_vet_TEN, main = "Sequence distribution plot", border = NA, with.legend = "right")
seqfplot(sts_vet_TEN, main = "Sequence frequency plot", with.legend = TRUE ,pbarw = TRUE)

tr_TEN <- seqtrate(sts_vet_TEN)
round(tr_TEN, 2)
 
```

### First 10 years clusters
```{r include = TRUE, echo = TRUE, warning = FALSE, fig.height=10, message = FALSE}
plot_clusters <- function(data, n) {
  cl1.n <- cutree(data, k = n)
  cl1.nfac <- factor(cl1.n, labels = paste("Type", 1:n))
  table(cl1.n)
  seqrplot(sts_vet_TEN, diss = dis_om_TEN, group = cl1.nfac, border = NA)
  seqdplot(sts_vet_TEN, group = cl1.nfac, border = NA)
  seqfplot(sts_vet_TEN, group = cl1.nfac, border = NA)
  seqmtplot(sts_vet_TEN, group = cl1.nfac, border = NA)
  seqHtplot(sts_vet_TEN, group = cl1.nfac, border = NA)
}

par(mar = (c(2,2,2,2)))
plot_clusters(clusterward_TEN, 3)
plot_clusters(clusterward_TEN, 4)
plot_clusters(clusterward_TEN, 5)
plot_clusters(clusterward_TEN, 6)
plot_clusters(clusterward_TEN, 7)
plot_clusters(clusterward_TEN, 8)
```
