==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/08_baseline_onet55_clustering.Rmd <==
---
title: "08 Post-Military Career Clustering"
author:
  "Crystal Zang"
output: 
  rmarkdown::html_document:
    toc: true
    toc_float: true
    highlight: tango
editor_options: 
  chunk_output_type: console
weight: 1
draft: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
for (pkg in c("tidyverse",  "data.table", "R.utils", "maditr", "stringr", "stringi", "dplyr", "ggplot2", "lubridate", "gt", "DataExplorer", "TraMineR", "TraMineRextras", "cluster")) {
  library(pkg, character.only = TRUE)
}

uva_color_palette <- 
c("#232D4B", #space cadet
  "#2C4F6B", #indigo dye
  "#0E879C", #blue munsell
  "#60999A", #cadet blue
  "#D1E0BF", #tea green
  "#D9E12B", #pear
  "#E6CE3A", #citrine
  "#E6A01D", #marigold
  "#E57200" #princeton orange
)
```


```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
bg_vet_job <-read_csv("~/git/dspg20career/data/04_bg_vet_job.csv")%>%
  mutate(is_onet55 = if_else(str_detect(onet, "55-[0-9][0-9][0-9][0-9].[0-9][0-9]"), T, F))%>%
  select(id,onet,is_onet55, onet_job_zone, startdate, enddate, start_year, end_year, onet_title, date_enter_job_market, job_duration_day)%>%
  mutate(year_enter_job_market = year(date_enter_job_market))

#how many onet55 job each person have
df <- bg_vet_job%>%
  group_by(id)%>%
  summarize(ttl_onet55_job = sum(is_onet55), ttl_job = n())

table(df$ttl_onet55_job)
```
Among veteran population, `r table(df$ttl_onet55_job)[1]/nrow(df)*100`% (N=`r table(df$ttl_onet55_job)[1]`) does not have a valid onet 55 job (missing date);     
`r table(df$ttl_onet55_job)[2]/nrow(df)*100`% (N=`r table(df$ttl_onet55_job)[2]`)have one valid onet 55 job;    
`r table(df$ttl_onet55_job)[3]/nrow(df)*100`% (N=`r table(df$ttl_onet55_job)[3]`) have 2 valid onet 55 job;   
`r table(df$ttl_onet55_job)[4]/nrow(df)*100`% (N=`r table(df$ttl_onet55_job)[4]`) have 4 valid onet 55 job.   


Problem: All 55-ONET jobs are missing ONET job zone 
The most frequent 55-ONET job 55-1019.00 is "Military Officer Special and Tactical Operations Leaders" (All military officer special and tactical operations leaders not listed separately).   

The second frequent 55-ONET job is 55-3019.00 "Military Enlisted Tactical Operations and Air/Weapons Specialists and Crew Members" (All military enlisted tactical operations and air/weapons specialists and crew members not listed separately)

ONET claims that they military occupation does not collect data on military occupations.
```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
bg_vet_job%>%
  filter(is_onet55)%>%
  group_by(onet, onet_job_zone)%>%
  summarize(N=n())%>%
  arrange(desc(N))
#make sure we get the end date of the last onet55 job
```

We have 3412 veterans in our sample who have valid onet55 job (with start and end date). The average year they end their last military job (ONET55 job) is 2000, with standard deviation of 10.
```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE,  fig.height=5, fig.width = 5}
vet_endmilitary <- bg_vet_job%>%
  mutate(date_end_onet55 = if_else(is_onet55==T, enddate, as.Date(NA)))%>%
  filter(!is.na(date_end_onet55))%>%  #exluce people who don't have valid onet55 code
  select(id, date_end_onet55) %>%
  #keep the latest onet55 job
  group_by(id)%>%   
  arrange(desc(date_end_onet55))%>%
  group_by(id)%>%
  distinct(id, .keep_all = TRUE)

# summary(bg_vet_job$year_end_onet55)
# sd(bg_vet_job$year_end_onet55)

ggplot(vet_endmilitary, aes(date_end_onet55))+
  geom_histogram()+
  labs(x = "year when the last ONET 55 job end")

bg_vet_job  <- inner_join(bg_vet_job, vet_endmilitary, by = "id")
```

# Sequence Analysis
# I. Complete sequence after the finishing the last *ONET 55 job
```{r, fig.height = 3, fig.width = 9}
bg_vet_job_seq <- bg_vet_job %>%
  select(id, end_year, onet_job_zone, startdate, enddate, job_duration_day, date_end_onet55, year_enter_job_market)%>%
  filter(startdate > date_end_onet55)  %>% #find jobs that came after the date ended onet55 job
  mutate(start_year = year(startdate))

ggplot(bg_vet_job_seq)+
    geom_segment(aes(x=as.character(start_year),
                     xend=as.character(end_year),
                     y=id,
                     yend=id), size=0.05,
                 color="purple")+
  theme_classic()+
  labs(x = "year", title = "Time in Work Force-Veteran")

```

```{r, fig.height = 5, fig.width = 9}
bg_vet_job_seq <- bg_vet_job_seq %>%
  mutate(start_year = start_year - year_enter_job_market + 1)%>%
  mutate(end_year = end_year - year_enter_job_market + 1)%>%
  select(id, start_year, end_year, onet_job_zone)%>%
  group_by(id)%>%
  arrange(desc(onet_job_zone))%>%
  group_by(id)%>%
  distinct(id, start_year, end_year, .keep_all = TRUE)

table(bg_vet_job_seq$onet_job_zone)
length(unique(bg_vet_job_seq$id))  #confirm sample size match

bg_vet_job_seq <- as.matrix(bg_vet_job_seq)
bg_vet_job_seq <- as.data.frame(bg_vet_job_seq)

min(bg_vet_job_seq$start_year)
max(bg_vet_job_seq$start_year)

sts_vet <- seqformat(bg_vet_job_seq, from = "SPELL", to = "STS",
                     id = "id",  begin = "start_year", end = "end_year", 
                     status = "onet_job_zone", process = FALSE)



# Here we are renaming columns to be in format "yn" (year in the job market)
names(sts_vet) <- paste0("y", 1:ncol(sts_vet))

nrow(sts_vet)  #total number of samples

#delete missing values at front and end
vet.seq <- seqdef(sts_vet, left="DEL", gaps="NA", right="DEL")

# vet.seq <- seqdef(sts_vet, left="DEL", gaps="NA", right="DEL", cpal = c( uva_color_palette[4],  uva_color_palette[3], uva_color_palette[2],uva_color_palette[2], uva_color_palette[8], "#CBBEB5"), with.missing=TRUE)

class(vet.seq)

#example of three sequences
vet.seq[1, ]
vet.seq[2, ]
vet.seq[100, ]

#substitution-cost matrix
#method="CONSTANT"/"TRATE"
cost_matrix_trate <- seqsubm(vet.seq, method = "TRATE", with.missing = TRUE)
cost_matrix_trate <- round(cost_matrix_trate, 2)
cost_matrix_trate

cost_matrix_constant <- seqsubm(vet.seq, method = "CONSTANT", with.missing = TRUE)
```

### I.a. Constant cost matrix
```{r}
# vet.seq.OM <- seqdist(vet.seq, method = "OM", indel = 3, sm = cost_matrix_constant, with.missing = TRUE)
# clusterward <- agnes(vet.seq.OM, diss = TRUE, method = "ward")
# saveRDS(clusterward, file = "data/clusterward_onet55_complete_constant_matrix.rds")

clusterward <- readRDS(file = "~/git/dspg20career/data/clusterward_onet55_complete_constant_cost_matrix.rds") 

#dendrogram 
plot(clusterward, which.plots =2)

```

#### I.a.i. Three clusters
```{r, fig.height = 8, fig.width = 8}
cluster3 <- cutree(clusterward, k=3)
cluster3 <- factor(cluster3, labels = c("Type 1",  "Type 2", "Type 3"))
table(cluster3)

#longitudinal plot
seqfplot(vet.seq, group = cluster3, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster3)
```

#### I.a.ii. Five clusters
```{r, fig.height = 8, fig.width = 8}
cluster5 <- cutree(clusterward, k=5)
cluster5 <- factor(cluster5, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5"))
table(cluster5)

#longitudinal plot
seqfplot(vet.seq, group = cluster5, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster5)
```


#### I.a.iii. Six clusters
```{r, fig.height = 8, fig.width = 8}
cluster6 <- cutree(clusterward, k=6)
cluster6 <- factor(cluster6, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5", "Type 6"))
table(cluster6)

#longitudinal plot
seqfplot(vet.seq, group = cluster6, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster6)
```

### I.b. Trate cost matrix
#### I.b.i. Five clusters
```{r, fig.height = 8, fig.width = 8}
# vet.seq.OM <- seqdist(vet.seq, method = "OM", indel = 3, sm = cost_matrix_trate, with.missing = TRUE)
# clusterward <- agnes(vet.seq.OM, diss = TRUE, method = "ward")
# saveRDS(clusterward, file = "data/clusterward_onet55_complete_trate_matrix.rds")

clusterward <- readRDS(file = "~/git/dspg20career/data/clusterward_onet55_complete_trate_cost_matrix.rds") 

#dendrogram 
plot(clusterward, which.plots =2)
```

```{r, fig.height = 8, fig.width = 8}
cluster5 <- cutree(clusterward, k=5)
cluster5 <- factor(cluster5, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5"))
table(cluster5)

#longitudinal plot
seqfplot(vet.seq, group = cluster5, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster5)
```


# II. Partial sequence after the 10 years finishing the last *ONET 55 job
```{r}
bg_vet_job_seq <- bg_vet_job %>%
  select(id, end_year, onet_job_zone, startdate, enddate, job_duration_day, date_end_onet55, year_enter_job_market)%>%
  filter(startdate > date_end_onet55)  %>% #find jobs that came after the date ended onet55 job
  mutate(start_year = year(startdate))%>%
  mutate(start_year = start_year - year_enter_job_market + 1)%>%
  mutate(end_year = end_year - year_enter_job_market + 1)

hist(bg_vet_job_seq$start_year)

#obtain jobs that appear after 10 years they exit military 
bg_vet_job_seq <- bg_vet_job_seq%>%
  filter(start_year > 10)%>%
  select(id, start_year, end_year, onet_job_zone)%>%
  group_by(id)%>%
  arrange(desc(onet_job_zone))%>%
  group_by(id)%>%
  distinct(id, start_year, end_year, .keep_all = TRUE)

table(bg_vet_job_seq$onet_job_zone)
length(unique(bg_vet_job_seq$id))  #confirm sample size match

bg_vet_job_seq <- as.matrix(bg_vet_job_seq)
bg_vet_job_seq <- as.data.frame(bg_vet_job_seq)

min(bg_vet_job_seq$start_year)
max(bg_vet_job_seq$start_year)

sts_vet <- seqformat(bg_vet_job_seq, from = "SPELL", to = "STS",
                     id = "id",  begin = "start_year", end = "end_year", 
                     status = "onet_job_zone", process = FALSE)

# Here we are renaming columns to be in format "yn" (year in the job market)
names(sts_vet) <- paste0("y", 1:ncol(sts_vet))

nrow(sts_vet)  #total number of samples

#delete missing values at front and end
vet.seq <- seqdef(sts_vet, left="DEL", gaps="NA", right="DEL")
class(vet.seq)

#example of three sequences
vet.seq[1, ]
vet.seq[2, ]
vet.seq[100, ]

#substitution-cost matrix
#method="CONSTANT"/"TRATE"
cost_matrix_trate <- seqsubm(vet.seq, method = "TRATE", with.missing = TRUE)
cost_matrix_trate <- round(cost_matrix_trate, 2)
cost_matrix_trate

cost_matrix_constant <- seqsubm(vet.seq, method = "CONSTANT", with.missing = TRUE)
```

### II.a. Constant cost matrix
```{r}
# vet.seq.OM <- seqdist(vet.seq, method = "OM", indel = 3, sm = cost_matrix_constant, with.missing = TRUE)
# clusterward <- agnes(vet.seq.OM, diss = TRUE, method = "ward")
# saveRDS(clusterward, file = "data/clusterward_onet55_after10yrs_constant_matrix.rds")

clusterward <- readRDS(file = "~/git/dspg20career/data/clusterward_onet55_after10yrs_constant_matrix.rds")

#dendrogram 
plot(clusterward, which.plots =2)
```

#### II.a.i. Three clusters
```{r, fig.height = 8, fig.width = 8}
cluster3 <- cutree(clusterward, k=3)
cluster3 <- factor(cluster3, labels = c("Type 1",  "Type 2", "Type 3"))
table(cluster3)

#longitudinal plot
seqfplot(vet.seq, group = cluster3, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster3)
```

#### II.a.ii. Four cluster
```{r, fig.height = 8, fig.width = 8}
cluster4 <- cutree(clusterward, k=4)
cluster4 <- factor(cluster4, labels = c("Type 1",  "Type 2", "Type 3", "Type 4"))
table(cluster4)

#longitudinal plot
seqfplot(vet.seq, group = cluster4, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster4)
```

#### II.a.iii. Five cluster
```{r, fig.height = 8, fig.width = 8}
cluster5 <- cutree(clusterward, k=5)
cluster5 <- factor(cluster5, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5"))
table(cluster5)

#longitudinal plot
seqfplot(vet.seq, group = cluster5, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster5)
```


#### II.a.iv. Six cluster
```{r, fig.height = 8, fig.width = 8}
cluster6 <- cutree(clusterward, k=6)
cluster6 <- factor(cluster6, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5", "Type 6"))
table(cluster6)

#longitudinal plot
seqfplot(vet.seq, group = cluster6, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster6)
```

#### II.a.v. Six cluster
```{r, fig.height = 8, fig.width = 8}
cluster7 <- cutree(clusterward, k=7)
cluster7 <- factor(cluster7, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5", "Type 6", "Type 7"))
table(cluster7)

#longitudinal plot
seqfplot(vet.seq, group = cluster7, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster7)
```

# III. Partial sequence of 10 years after finishing the last *ONET 55 job
```{r}
bg_vet_job_seq <- bg_vet_job %>%
  select(id, end_year, onet_job_zone, startdate, enddate, job_duration_day, date_end_onet55, year_enter_job_market)%>%
  filter(startdate > date_end_onet55)  %>% #find jobs that came after the date ended onet55 job
  mutate(start_year = year(startdate))%>%
  mutate(start_year = start_year - year_enter_job_market + 1)%>%
  mutate(end_year = end_year - year_enter_job_market + 1)

bg_vet_job_seq <- bg_vet_job_seq%>%
  select(id, start_year, end_year, onet_job_zone)%>%
  group_by(id)%>%
  arrange(desc(onet_job_zone))%>%
  group_by(id)%>%
  distinct(id, start_year, end_year, .keep_all = TRUE)

table(bg_vet_job_seq$onet_job_zone)
length(unique(bg_vet_job_seq$id))  #confirm sample size match

bg_vet_job_seq <- as.matrix(bg_vet_job_seq)
bg_vet_job_seq <- as.data.frame(bg_vet_job_seq)

min(bg_vet_job_seq$start_year)
max(bg_vet_job_seq$start_year)

sts_vet <- seqformat(bg_vet_job_seq, from = "SPELL", to = "STS",
                     id = "id",  begin = "start_year", end = "end_year", 
                     status = "onet_job_zone", process = FALSE)

class(sts_vet)
#obtain jobs that appear after 10 years they exit military 
sts_vet <- sts_vet[, 1:10]


# Here we are renaming columns to be in format "yn" (year in the job market)
names(sts_vet) <- paste0("y", 1:ncol(sts_vet))

nrow(sts_vet)  #total number of samples

#delete missing values at front and end
vet.seq <- seqdef(sts_vet, left="NA", gaps="NA", right="DEL")
class(vet.seq)

#example of three sequences
vet.seq[1, ]
vet.seq[2, ]
vet.seq[100, ]

#substitution-cost matrix
#method="CONSTANT"/"TRATE"
cost_matrix_trate <- seqsubm(vet.seq, method = "TRATE", with.missing = TRUE)
cost_matrix_trate <- round(cost_matrix_trate, 2)
cost_matrix_trate

cost_matrix_constant <- seqsubm(vet.seq, method = "CONSTANT", with.missing = TRUE)
```

### II.a. Constant cost matrix
```{r, fig.height = 10, fig.width = 8}
# vet.seq.OM <- seqdist(vet.seq, method = "OM", indel = 3, sm = cost_matrix_constant, with.missing = TRUE)
# clusterward <- agnes(vet.seq.OM, diss = TRUE, method = "ward")
# saveRDS(clusterward, file = "data/clusterward_onet55_10yrs_constant_matrix.rds")

clusterward <- readRDS(file = "~/git/dspg20career/data/clusterward_onet55_10yrs_constant_matrix.rds")

#dendrogram 
plot(clusterward, which.plots =2)
```

#### II.a.i. Three clusters
```{r, fig.height = 8, fig.width = 8}
cluster3 <- cutree(clusterward, k=3)
cluster3 <- factor(cluster3, labels = c("Type 1",  "Type 2", "Type 3"))
table(cluster3)

#longitudinal plot
seqfplot(vet.seq, group = cluster3, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster3)
```

#### II.a.ii. Four clusters
```{r, fig.height = 8, fig.width = 8}
cluster4 <- cutree(clusterward, k=4)
cluster4 <- factor(cluster4, labels = c("Type 1",  "Type 2", "Type 3", "Type 4"))
table(cluster4)#### II.a.i. Three clusters
```

```{r, fig.height = 8, fig.width = 8}
cluster3 <- cutree(clusterward, k=3)
cluster3 <- factor(cluster3, labels = c("Type 1",  "Type 2", "Type 3"))
table(cluster3)

#longitudinal plot
seqfplot(vet.seq, group = cluster3, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster3)
```

#longitudinal plot
seqfplot(vet.seq, group = cluster4, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster4)
```

#### II.a.iii. Five clusters
```{r, fig.height = 10, fig.width = 8}
cluster5 <- cutree(clusterward, k=5)
cluster5 <- factor(cluster5, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type5"))
table(cluster5)

#longitudinal plot
seqfplot(vet.seq, group = cluster5, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster5)
```

#### II.a.iv. Six clusters
```{r, fig.height = 10, fig.width = 8}
cluster6 <- cutree(clusterward, k=6)
cluster6 <- factor(cluster6, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5", "Type 6"))
table(cluster6)

#longitudinal plot
seqfplot(vet.seq, group = cluster6, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster6)
```


#### II.a.v. Seven clusters
```{r, fig.height = 10, fig.width = 8}
cluster7 <- cutree(clusterward, k=7)
cluster7 <- factor(cluster7, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5", "Type 6", "Type 7"))
table(cluster7)

#longitudinal plot
seqfplot(vet.seq, group = cluster7, pbarw = T)

#another cluster plot
seqmtplot(vet.seq, group = cluster7)
```
