==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/10_exploring_sequences.Rmd <==
---
title: "10_exploring_sequences"
output: html_document
---

Exploring sequence object through graphics. Here we explore both veterans entire career sequence and the first ten years after military exit.

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
for (pkg in c("tidyverse",  "data.table", "R.utils", "maditr", "stringr", "stringi", "dplyr", "ggplot2", "lubridate", "gt", "DataExplorer", "TraMineR", "TraMineRextras", "cluster", "viridis")) {
  library(pkg, character.only = TRUE)
}

uva_color_palette <- 
c("#232D4B", #space cadet
  "#2C4F6B", #indigo dye
  "#0E879C", #blue munsell
  "#60999A", #cadet blue
  "#D1E0BF", #tea green
  "#D9E12B", #pear
  "#E6CE3A", #citrine
  "#E6A01D", #marigold
  "#E57200" #princeton orange
)
```

```{r}
## Creating sequence object for entire career ------

bg_vet_job <- read_csv("~/git/DSPG2020/career/data/04_bg_vet_job.csv")
bg_vet_demographic <- read_csv("~/git/DSPG2020/career/data/02_bg_vet_demographic.csv")

# We are choosing years as our unit of analysis because months are unreliable. We are also taking the highest O*NET job zone for a given year to deal with overlapping jobs.
bg_vet_job_seq <- bg_vet_job %>%
  mutate(startyear = year(startdate), endyear = year(enddate)) %>%
  select("id", "startyear", "endyear", "onet_job_zone")%>%
  group_by(id)%>%
  arrange(desc(onet_job_zone))%>%
  group_by(id)%>%
  distinct(id, startyear, endyear, .keep_all = TRUE)

# We need an auxillary table with the year of entering the job market to align the sequences
bg_vet_job_first <- bg_vet_job_seq %>% 
  select("id", "startyear") %>% group_by(id) %>% transmute(enter = min(startyear)) %>% distinct() %>% ungroup()

# The seqformat() function does not like any data that has been previously grouped. Here we are "resetting" the data so that it will pass through the function.
bg_vet_job_seq <- as.matrix(bg_vet_job_seq)
bg_vet_job_seq <- as.data.frame(bg_vet_job_seq)
bg_vet_job_first <- as.matrix(bg_vet_job_first)
bg_vet_job_first <- as.data.frame(bg_vet_job_first)

# The input for the function is the prepared sequence table. The data is in format SPELL and we are transforming to format STS. By setting process = TRUE we can align the sequences using the prepared auxillary table.
sts_vet <- seqformat(bg_vet_job_seq, from = "SPELL", to = "STS",
                     id = "id",  begin = "startyear", end = "endyear", 
                     status = "onet_job_zone", process = TRUE,
                     pdata = bg_vet_job_first, pvar = c("id", "enter"))
# Here we are renaming columns to be in format "yn" (year in the job market)
names(sts_vet) <- paste0(1:100)

# Writing the final table as a csv
#write.csv(sts_vet, "sts_vet_55.csv")

sts_vet <- seqdef(sts_vet, cpal = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "red", "#CBBEB5"), left="DEL", gaps="unemployed", right="DEL",
labels = c("Missing job zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Unemployment"))

```

```{r}
## Sequence graphics for entire career using functions in TraMineR ------

# Seqmtplot displays the mean time spent in each state
seqmtplot(sts_vet, main = "Mean time distribution plot", ylim = c(0, 10), border = NA, with.legend= FALSE)

# Seqdplot represents the sequence of the cross-sectional state frequencies by position
seqdplot(sts_vet, main = "Sequence distribution plot", border = NA, with.legend = "right")

# Seqfplot displays the most frequent sequences
seqfplot(sts_vet, main = "Sequence frequency plot", with.legend = TRUE ,pbarw = TRUE)

seqHtplot(sts_vet, main = "Entropy index plot", with.legend = TRUE)


#This indicator is called the entropy index Billari [2001] It equals 0 when all cases are in the same state (it is thus easyto predict in which state an individual is) It is maximum when the cases are equally distributed betweenthe states of the alphabet (it is thus hard to predict in whichstate an individual is)

# Transition matrix, we use ggplot2 to create this
tr <- seqtrate(sts_vet)
round(tr, 2)

transition.matrix.df <- as.data.frame(tr)
transition.matrix.df$startState <- row.names(transition.matrix.df)
transition.matrix.df <- melt(transition.matrix.df)
colnames(transition.matrix.df)[colnames(transition.matrix.df) == "variable"] <- "endState"
                                        
ggplot(data = transition.matrix.df, aes(x = endState, y = factor(startState,        levels = rev(levels(factor(startState)))), fill = value)) + 
  geom_tile() +    
  geom_text(aes(label=round(value,3))) +  
  scale_x_discrete(position = "top") +  
  scale_fill_gradient(low = "white", high = "#E57200") +
  theme(legend.position="bottom", panel.background = element_blank(),         plot.title = element_text(face = "bold")) +  
  labs(fill = "Transition Probability",        
       title = "Transition Matrix for Veterans",     
       y = "Current Job",       
       x = "Next Job")

sts_vet_demographic <- left_join(sts_vet %>% mutate(id = as.numeric(rownames(sts_vet))), bg_vet_demographic, by = "id")

sts_vet_demographic <- seqdef(sts_vet_demographic, 1:100, left="DEL", gaps="unemployed", right="DEL")

```


```{r}
## Sequence graphics for entire career using ggplot2 -------

statd <- seqstatd(sts_vet)
states <- statd$Frequencies
states <- as.data.frame(states)
names(states) <- paste0(1:100)
states <- states %>% rownames_to_column(var = "state")
states <- melt(states, id = "state")

v_states <- statd$ValidStates
v_states <- as.data.frame(v_states)
v_states <- v_states %>% rownames_to_column(var = "variable")

meant <- seqmeant(sts_vet)
meant <- as.data.frame(meant)
meant <- meant %>% rownames_to_column(var = "state")

lengths <- seqlength(sts_vet)
lengths <- as.data.frame(lengths)
lengths <- lengths %>% rownames_to_column(var = "id")
length_sum <- quantile(lengths$Length)
length_sum <- as.data.frame(length_sum)
length_sum <- length_sum %>% rownames_to_column(var = "measure")
summary(lengths)

library(scales)
ggplot() + 
  geom_area(data = states, aes(x = as.numeric(variable), y = value, fill = state, position = "fill")) +
  geom_line(data = v_states, aes(x = as.numeric(variable), y = v_states/8068),
            size = 2) +
  scale_y_continuous(name = "Share of careers in a job state",
                     sec.axis = sec_axis(trans = ~.*(8068), name = "Number of careers in year of career")) +
  xlim(0,78) +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2], uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "red", "#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Unemployed")) +
  xlab("Year of career") +
  labs(title = "Visualizing career paths of veterans: Job states over time",
       fill = "Job state") +
  theme_classic() +
  theme(legend.position = "bottom")

ggplot(data = lengths, aes(x = Length)) +
  geom_histogram(binwidth = 5, colour="black", fill = "white") +
  geom_vline(data = length_sum, aes(xintercept = 20,
                                    color="Median"), linetype="solid", size=1) +
  geom_vline(data = length_sum, aes(xintercept = 13,
                                    color= "1st Qu."), linetype="solid", size=1) +
  geom_vline(data = length_sum, aes(xintercept = 29,
                                    color="3rd Qu."), linetype="solid", size=1) +
  geom_vline(data = length_sum, aes(xintercept = 1,
                                    color="Min"), linetype="solid", size=1) +
  geom_vline(data = length_sum, aes(xintercept = 86,
                                    color="Max"), linetype="solid", size=1) +
  xlim(1,86) +
  labs(title = "Lengths of career sequences for veteran's whole careers",
       caption = "Data: BGT Resumes") +
  xlab("Sequence length (years)") +
  ylab("Count of sequences") +
  #  annotate("text", x=74, y=364, 
  #           label=paste("Min. =", length_sum[1,2],"\n Max. = ", length_sum[5,2]), size=3) +
  theme_classic() +
  theme(legend.position=c(.8,.5)) +
  scale_color_viridis_d(name = "Values", limits = c("Min", "1st Qu.", "Median", "3rd Qu.", "Max"))

ggplot(data = meant, aes(x = state, y = Mean, fill = state)) +
  geom_col() +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2], uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "red", "#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Unemployed")) +
  scale_x_discrete(name = "Job state", labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Unemployed")) +
  labs(title = "Mean time spent in each state for veteran's whole career sequence",
       fill = "Job state",
       caption = "Data: BGT Resumes") +
  ylab("Mean time (years)") +
  theme_classic() +
  theme(legend.position = "none",
        axis.text.x = element_text(angle = 40, hjust = 1))

ggplot() + 
  geom_area(data = states, aes(x = as.numeric(variable), y = value, fill = state, position = "fill")) +
  geom_line(data = v_states, aes(x = as.numeric(variable), y = v_states/8068),
            size = 2) +
  scale_y_continuous(name = "Share of careers in a job state",
                     sec.axis = sec_axis(trans = ~.*(8068), name = "Number of careers in year of career")) +
  xlim(0,78) +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2], uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "red", "#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Unemployed")) +
  xlab("Year of career") +
  labs(title = "Job states over time for veteran's whole career sequence",
       fill = "Job state",
       caption = "Data: BGT Resumes") +
  theme_classic() +
  theme(legend.position = "bottom")
```

```{r}
# Creating sequence object for first ten years after military exit -----

# Remove objects from environment
remove(list = ls())

uva_color_palette <- c("#232D4B", #space cadet
  "#2C4F6B", #indigo dye
  "#0E879C", #blue munsell
  "#60999A", #cadet blue
  "#D1E0BF", #tea green
  "#D9E12B", #pear
  "#E6CE3A", #citrine
  "#E6A01D", #marigold
  "#E57200" #princeton orange
)

## Code from Crystal's 11_baseline_onet55_clustering_updated.Rmd and results-section-clustering.Rmd
source("~/git/DSPG2020/career/src/burningglass/clean_post_military_job.R", echo = T, prompt.echo = "", spaced = F)
 
#read in vet job data
bg_vet_job <-read_csv("~/git/DSPG2020/career/data/04_bg_vet_job.csv")
 
#perform the cleaning function
bg_vet_job_seq <- clean_post_military(bg_vet_job)

bg_vet_job_seq <- as.matrix(bg_vet_job_seq)
bg_vet_job_seq <- as.data.frame(bg_vet_job_seq)

# Convert sequence data from SPELL format to STS format
sts_vet <- seqformat(bg_vet_job_seq, from = "SPELL", to = "STS",
                     id = "id",  begin = "start_year", end = "end_year", 
                     status = "onet_job_zone", process = FALSE)

#obtain jobs that appear after 10 years they exit military 
sts_vet <- sts_vet[, 1:10]

# Here we are renaming columns to be in format "yn" (year in the job market)
names(sts_vet) <- paste0("y", 1:ncol(sts_vet))

#create sequence object
#assign left NA as transitional unemployment, gaps as unemployment, and right NA as unemployment or exit
#assign colors to different categories

vet.seq <- seqdef(sts_vet, 
                  left="Military Transitional Unemployment", gaps="Civilian Unemployment", right="Retirement", 
                  cpal = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#CBBEB5", "#CBBEB5","#CBBEB5")
                  )

clusterward <- readRDS(file = "~/git/DSPG2020/career/data/clusterward_onet55_10yrs_normalized_matrix.rds")
cluster8 <- cutree(clusterward, k=8)
cluster8 <- factor(cluster8, labels = c("Type 1",  "Type 2", "Type 3", "Type 4", "Type 5", "Type 6", "Type 7", "Type 8"))

```

```{r}
## Sequence graphics for first ten years after military exit using ggplot2 ------

statd <- seqstatd(vet.seq)
states <- statd$Frequencies
states <- as.data.frame(states)
names(states) <- paste0(1:10)
states <- states %>% rownames_to_column(var = "state")
states <- melt(states, id = "state")

v_states <- statd$ValidStates
v_states <- as.data.frame(v_states)
v_states <- v_states %>% rownames_to_column(var = "variable")

seqdplot(vet.seq)
seqmtplot(vet.seq)

ggplot() + 
  geom_area(data = states, aes(x = as.numeric(variable), y = value, fill = state, position = "fill")) +
 # geom_line(data = v_states, aes(x = as.numeric(variable), y = v_states/5185),
 #           size = 2) +
  scale_y_continuous(name = "Share of careers in a job state") +
  xlim(0,10) +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2], uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "red", "gray", "#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Transitional unemployed", "Unemployed")) +
  xlab("Year of career") +
  labs(title = "Visualizing career paths of veterans: Job states over time",
       fill = "Job state") +
  theme_classic() +
  theme(legend.position = "bottom")

meant <- seqmeant(vet.seq)
meant <- as.data.frame(meant)
meant <- meant %>% rownames_to_column(var = "state")

ggplot(data = meant, aes(x = state, y = Mean, fill = state)) +
  geom_col() +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#CBBEB5", "#CBBEB5","#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Transition unemployed", "Unemployed", "Unemployed/Exit")) +
  scale_x_discrete(name = "Job state", labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military", "Transition unemployed", "Unemployed", "Unemployed/Exit")) +
  labs(title = "Mean time spent in each state",
       fill = "Job state") +
  ylab("Mean time (years)") +
  theme_classic() +
  theme(legend.position = "none")

ggplot(data = meant, aes(x = state, y = Mean, fill = state)) +
  geom_col() +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#CBBEB5", "#CBBEB5","#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Transition unemployed", "Unemployed", "Unemployed/Exit")) +
  scale_x_discrete(name = "Job state", labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Transition unemployed", "Unemployed", "Unemployed/Exit")) +
  labs(title = "Mean time spent in each state for first ten years \n after veteran's last military job",
       fill = "Job state",
       caption = "Data: BGT Resumes") +
  ylab("Mean time (years)") +
  theme_classic() +
  theme(legend.position = "none", axis.text.x = element_text(angle = 40, hjust = 1))


p <- ggplot() + 
  geom_area(data = states, aes(x = as.numeric(variable), y = value, fill = state, position = "fill")) +
  # geom_line(data = v_states, aes(x = as.numeric(variable), y = v_states/5185),
  #           size = 2) +
  scale_y_continuous(name = "Share of careers in a job state") +
  xlim(0,10) +
  scale_fill_manual(values = c("#CBBEB5", uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#DCDCDC", "#C0C0C0", "#696969"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military transition unemployed", "Civilian unemployed", "Retired")) +
  xlab("Year of career") +
  labs(title = "Job states over time for first ten years \n after veteran's last military job",
       fill = "Job state",
       caption = "Data: BGT Resumes") +
  theme_classic() +
  theme(legend.position = "bottom")

p +
  scale_x_continuous(breaks = pretty_breaks(4))

library(scales)
p <- ggplot() + 
  geom_area(data = states, aes(x = as.numeric(variable), y = value, fill = state, position = "fill")) +
  # geom_line(data = v_states, aes(x = as.numeric(variable), y = v_states/5185),
  #           size = 2) +
  #scale_y_continuous(name = "Share of careers in a job state") +
  xlim(0,10) +
  scale_fill_manual(values = c("#CBBEB5", uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#DCDCDC", "#C0C0C0", "#696969"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military transition unemployed", "Civilian unemployed", "Retired")) +
  xlab("Year of career") +
  labs(fill = "Job state") +
  ylab("Number of careers in year of career") +
  theme_classic() +
  theme(legend.position = "bottom",
        axis.title.y=element_blank(),
        axis.text.y=element_blank())


dten <- p +
  scale_x_continuous(breaks = pretty_breaks(4))
 


s <- ggplot(data = meant, aes(x = state, y = Mean, fill = state)) +
  geom_col() +
  scale_fill_manual(values = c("#CBBEB5",  uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#CBBEB5", "#CBBEB5","#CBBEB5"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Transition unemployed", "Unemployed", "Unemployed/Exit")) +
  scale_x_discrete(name = "Job state", labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military transition unemployed", "Civilian unemployed", "Retired")) +
  labs(title = "Mean time spent in each state for first ten years after veteran's last military job",
       fill = "Job state",
       caption = "Data: BGT Resumes") +
  ylab("Mean time (years)") +
  theme_classic() +
  theme(legend.position = "none", axis.text.x = element_text(angle = 40, hjust = 1))


dten <- s +
  scale_y_continuous(breaks = c(0,1,2,3)) +
  ylim(0,3)


library(grid)
library(gridExtra)
#grid.arrange(
#  dall,
#  dten,
#  nrow = 1,
#  #top = "Comparison of job states over time"
#  top = textGrob(
#    "Comparison of job states over time",
#    gp = gpar(fontface = 1, fontsize = 18)))

```

```{r fig.height=10}
## Rugplots for cluster membership using TraMineR an dggplot2 ------

#clust.labels <- c("Employment", "Higher Education", "Training", "Joblessness")
#mvad.cl4.factor <- factor(mvad.cl4, levels = 1:8)

dist.mostfreq <- seqdist(vet.seq, method = "LCS", refseq = 0)

# TraMineR graphic for sequences by cluster membership
seqIplot(vet.seq, border = NA, group = cluster8, sortv = dist.mostfreq,
         xlim = c(0,10))

# TraMineR graphic for sequences by highest degree earned
#seqIplot(sts_vet, border = NA, group = demo$degree_highest, sortv = dist.mostfreq,
#         xlim = c(0,60))

vet.seq <- vet.seq %>% rownames_to_column(var = "id")
vet.seq$cluster <- cluster8
vet.seq$dist <- dist.mostfreq


vetseq_long <- vet.seq %>% 
  gather(key = "Year", value = "state", 2:11, -id)

# Computing transitions if you want to order by transitions
vetseq_long <- vetseq_long %>%
  group_by(id) %>%
  mutate(transitions = n_distinct(state))

# If you want to show clusters as facets you need to reorder them to the order that Crystal created for the narrative
#vet.seq$cluster_f = factor(vet.seq$cluster, levels=c("Type 1", "Type 2", "Type 3", "Type 4", "Type 5", "Type 6", "Type 7", "Type 8"))

# Rugplot for clusters, filtered by cluster and sorted by distance
vetseq_long %>% 
  filter(cluster == "Type 3") %>%
  ggplot(aes(x = Year, y = reorder(id, dist), fill = state)) +
  geom_tile() +
  #facet_wrap(~cluster, scales = "free_y", ncol = 2) +
  theme_bw() +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    panel.grid = element_blank(),
    text = element_text(size=20)
  ) +
  scale_fill_manual(values = c("#CBBEB5", uva_color_palette[1], uva_color_palette[2],uva_color_palette[4],  uva_color_palette[6], uva_color_palette[9], "#DCDCDC", "#C0C0C0", "#696969"), labels = c("Missing zone", "Zone 1", "Zone 2", "Zone 3", "Zone 4", "Zone 5", "Military transition unemployed", "Civilian unemployed", "Retired")) +
  labs(x = "", y = "",
       title = "Cluster 5: Zone 5 cluster") +
  scale_x_discrete(name ="Years in career", 
                    limits=c("y1", "y2", "y3", "y4", "y5", "y6", "y7", "y8", "y9", "y10"),
                   labels = c(1:10)) +
  guides(fill = guide_legend(override.aes = list(size = 20)))
  

```
