==> 01-Program-Years/2020 (3 State)/Project Code/dspg20career/src/burningglass/12_tournament_models.Rmd <==
---
title: "11 Tournament Models"
author:
  "Maddie Pickens"
output: 
  rmarkdown::html_document:
    toc: true
    toc_float: true
    highlight: tango
editor_options: 
  chunk_output_type: console
weight: 1
draft: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE,
                      warning = FALSE
)

for (pkg in c("tidyverse",  "data.table", "R.utils", "maditr", "stringr", "stringi", "dplyr", "ggplot2", "lubridate", "gt", "DataExplorer", "TraMineR", "TraMineRextras", "cluster", "MASS")) {
  library(pkg, character.only = TRUE)
}

uva_color_palette <- 
c("#232D4B", #space cadet
  "#2C4F6B", #indigo dye
  "#0E879C", #blue munsell
  "#60999A", #cadet blue
  "#D1E0BF", #tea green
  "#D9E12B", #pear
  "#E6CE3A", #citrine
  "#E6A01D", #marigold
  "#E57200" #princeton orange
)
```

## R Markdown

```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
source("~/git/dspg20career/src/burningglass/clean_post_military_job.R", echo = T, prompt.echo = "", spaced = F)

bg_vet_job <-read_csv("~/git/dspg20career/data/04_bg_vet_job.csv")
bg_vet_job_seq <- clean_post_military(bg_vet_job)

bg_vet_job_seq <- as.matrix(bg_vet_job_seq)
bg_vet_job_seq <- as.data.frame(bg_vet_job_seq)

sts_vet <- seqformat(bg_vet_job_seq, from = "SPELL", to = "STS",
                     id = "id",  begin = "start_year", end = "end_year", 
                     status = "onet_job_zone", process = FALSE)



# Here we are renaming columns to be in format "yn" (year in the job market)
names(sts_vet) <- paste0("y", 1:ncol(sts_vet))

vet.seq <- seqdef(sts_vet, left="NA", gaps="NA", right="DEL")

# limit to first 5 years
vet.seq.5 <- vet.seq[,1:5]
vet.seq.5 <- as.data.frame(vet.seq.5)
names <- colnames(vet.seq.5)

vet.seq.5[, names] <- sapply(vet.seq.5[, names],function(x) str_replace_all(x,fixed("%"), "NA"))

vet.seq.5.numeric <- vet.seq.5

vet.seq.5.numeric[,names] = data.frame(sapply(vet.seq.5.numeric[,names], as.character))
vet.seq.5.numeric[,names] = data.frame(sapply(vet.seq.5.numeric[,names], as.numeric))


# limit to first 10 years
vet.seq.10 <- vet.seq[,1:10]
vet.seq.10 <- as.data.frame(vet.seq.10)
names <- colnames(vet.seq.10)

vet.seq.10[, names] <- sapply(vet.seq.10[, names],function(x) str_replace_all(x,fixed("%"), "NA"))

vet.seq.10.numeric <- vet.seq.10
vet.seq.10.numeric[,names] = data.frame(sapply(vet.seq.10.numeric[,names], as.character))
vet.seq.10.numeric[,names] = data.frame(sapply(vet.seq.10.numeric[,names], as.numeric))

# limit to first 20 years
vet.seq.20 <- vet.seq[,1:20]
vet.seq.20 <- as.data.frame(vet.seq.20)

vet.seq.20[, names] <- sapply(vet.seq.20[, names],function(x) str_replace_all(x,fixed("%"), "NA"))


vet.seq.20.numeric <- vet.seq.20
names <- colnames(vet.seq.20.numeric)
vet.seq.20.numeric[,names] = data.frame(sapply(vet.seq.20.numeric[,names], as.character))
vet.seq.20.numeric[,names] = data.frame(sapply(vet.seq.20.numeric[,names], as.numeric))
```

### Group the sequences

```{r}
vet.seq.10 <- vet.seq.10.numeric %>%
  mutate(promoted_first_period = ifelse(((y3 > y2) & !is.na(y3)),
                                        'Promoted First Period', 'Not Promoted First Period'),
        promoted_second_period = ifelse(((y5 > y4) & !is.na(y5)),
                                        'Promoted Second Period', 'Not Promoted Second Period'),
        promoted_third_period = ifelse(((y7 > y5) & !is.na(y7)),
                                       'Promoted Third Period', 'Not Promoted Third Period'),
        promoted_fourth_period = ifelse(((y9 > y5) & !is.na(y9)),
                                         'Promoted Fourth Period', 'Not Promoted Fourth Period'),
        promoted_fifth_period = ifelse(((y10 > y9) & !is.na(y10)),
                                       'Promoted Fifth Period', 'Not Promoted Fifth Period'),
        unemployed_first_period = ifelse((is.na(y1) & is.na(y2)), 
                                         'Unemployed First Period', 'Employed First Period'),
        employed_later_period = ifelse((!is.na(y3) & !is.na(y4) & !is.na(y5) & !is.na(y6) & !is.na(y7) & !is.na(y8)
                                          & !is.na(y9) & !is.na(y10)), 'Employed Y3-Y10','Unemployed Later Period'),
        unemployed_later_period = ifelse((is.na(y3) | is.na(y4) | is.na(y5) | is.na(y6) | is.na(y7) | is.na(y8)
                                          | is.na(y9) | is.na(y10)), 'Employed Y3-Y10','Unemployed Later Period'),
        finish_zone_1_2_unemp = ifelse((y10 == 1 | y10 == 2 | is.na(y10)),
                                   'Finish Unemployed or Zone 1 or 2', 'Finish Zone 3, 4, or 5'),
        finish_zone_4_5 = ifelse((y10 == 4 | y10 == 5), 'Finish Zone 4 or 5', 'Finish UE or Z1, 2, 3')
        
)
```

```{r}
# promoted first period, promoted second period
tblA = table(vet.seq.10$promoted_first_period, vet.seq.10$promoted_second_period)
tblA
chisq.test(tblA)
```

```{r}
# promoted first period, employed full later period
tblB = table(vet.seq.10$promoted_first_period, vet.seq.10$employed_later_period)
tblB
chisq.test(tblB)
```

```{r}
# unemployed first period, unemployed later period
tblC = table(vet.seq.10$unemployed_first_period, vet.seq.10$unemployed_later_period)
tblC
chisq.test(tblC)
```

```{r}
# unemployed first period, zone 1 or 2 or unemployed at end of 10 years
tblD = table(vet.seq.10$unemployed_first_period, vet.seq.10$finish_zone_1_2_unemp)
tblD
chisq.test(tblD)
```

```{r}
# promoted first period, zone 4 or 5 at end of 10 years
tblE = table(vet.seq.10$promoted_first_period, vet.seq.10$finish_zone_4_5)
tblE
chisq.test(tblE)
```



```{r}
# vet.seq.10.promote <- vet.seq.10 %>%
#   mutate(promotepd1 = ifelse(((as.numeric(y3) > as.numeric(y2)) & y3 != 'NA'), TRUE, FALSE),
#          promotedlater = ifelse(((as.numeric(y10) > as.numeric(y1) & y10 != 'NA'), TRUE, FALSE)),
#          
# 
# 
# 
# prm_pd1 <- vet.seq.10 %>%
#   filter((y3 > y2) & !is.na(y3))
# 
# not_prm_pd1 <- vet.seq.10 %>%
#   filter(!(as.numeric(y3) > as.numeric(y2)))
# 
# prm_pd2 <- vet.seq.10 %>%
#   filter((as.numeric(y5) > as.numeric(y4)) & y5 != 'NA')
# 
# not_prm_pd2 <- vet.seq.10 %>%
#   filter(!(as.numeric(y5) > as.numeric(y4)))
# 
# prm_pd3 <- vet.seq.10 %>%
#   filter((as.numeric(y7) > as.numeric(y6)) & y7 != 'NA')
# 
# not_prm_pd3 <- vet.seq.10 %>%
#   filter(!(as.numeric(y7) > as.numeric(y6)))
# 
# prm_pd4 <- vet.seq.10 %>%
#   filter((as.numeric(y9) > as.numeric(y8)) & y9 != 'NA')
# 
# not_prm_pd4 <- vet.seq.10 %>%
#   filter(!(as.numeric(y9) > as.numeric(y8)))
# 
# prm_pd5 <- vet.seq.10 %>%
#   filter((as.numeric(y10) > as.numeric(y9)) & y10 != 'NA')
# 
# not_prm_pd5 <- vet.seq.10 %>%
#   filter(!(as.numeric(y10) > as.numeric(y9)))
# 
# lvl3_yr5 <- vet.seq.10 %>%
#   filter((as.numeric(y5) >= 3) & y5 != 'NA')
# 
# notlvl3_yr5 <- vet.seq.10 %>%
#   filter(!(as.numeric(y5) >= 3))
# 
# prm_later_pd <- vet.seq.10 %>%
#   filter(((as.numeric(y10) > as.numeric(y9)) & y10 != 'NA') |
#          ((as.numeric(y9) > as.numeric(y8)) & y9 != 'NA') |
#          ((as.numeric(y7) > as.numeric(y6)) & y7 != 'NA') |
#          ((as.numeric(y5) > as.numeric(y4)) & y5 != 'NA'))
# 
# lv45_yr10 <- vet.seq.10 %>%
#   filter((as.numeric(y10) >=4) & y10 != 'NA')
# 
# notlv45_yr10 <- vet.seq.10 %>%
#   filter(!(as.numeric(y10) >=4))
# 
# prmpd1_notpd2 <- vet.seq.10 %>%
#     filter(((as.numeric(y3) > as.numeric(y2)) & y3 != 'NA') &
#   (!(as.numeric(y5) > as.numeric(y4)) & y5 != 'NA'))
# 
# prmpd1_andpd2 <- vet.seq.10 %>%
#   filter(((as.numeric(y3) > as.numeric(y2)) & y3 != 'NA') &
#   ((as.numeric(y5) > as.numeric(y4)) & y5 != 'NA'))
# 
# prm_pd1 <- vet.seq.10 %>%
#   filter((as.numeric(y3) > as.numeric(y2)) & y3 != 'NA')
# 
# not_prmpd1_orpd2 <- vet.seq.10 %>%
#   filter(!(as.numeric(y3) > as.numeric(y2)) &
#           !(as.numeric(y5) > as.numeric(y4)))
# 
# not_prmpd1_prmpd2 <- vet.seq.10 %>%
#   filter((as.numeric(y5) > as.numeric(y4)) & y5 != 'NA' & !(as.numeric(y3) > as.numeric(y2)))
```

```{r}
# nrow(prmpd1_andpd2)
# nrow(prmpd1_notpd2)
# nrow(not_prmpd1_orpd2)
# nrow(not_prmpd1_prmpd2)
# tblA <- matrix(c(5, 167, 140, 2121), nrow = 2, ncol = 2)
# 
# # chances of being promoted in second period higher for those in first
# chisq.test(tblA)
```

```{r}
#promoted first period and level 3 in 5 years
# prm_pd1_and_lvl3_yr5 <- vet.seq.10 %>%
#   filter(((as.numeric(y5) >= 3) & y5 != 'NA') & ((as.numeric(y3) > as.numeric(y2)) & y3 != 'NA'))
# nrow(prm_pd1_and_lvl3_yr5)
# 
# # promoted first period, not level 3 in 5 years
# prm_pd1_not_lvl3_yr5 <- vet.seq.10 %>%
#   filter((!(as.numeric(y5) >= 3) & y5 != 'NA') & ((as.numeric(y3) > as.numeric(y2)) & y3 != 'NA'))
# nrow(prm_pd1_not_lvl3_yr5)
# 
# # promoted later period, and level 3 in 5 years
# prm_later_and_lvl3_yr5 <- vet.seq.10 %>%
#   filter(((as.numeric(y5) >= 3) & y5 != 'NA') & ((as.numeric(y10) > as.numeric(y9)) & y10 != 'NA') |((as.numeric(y9) > as.numeric(y8)) & y9 != 'NA') |((as.numeric(y7) > as.numeric(y6)) & y7 != 'NA') |((as.numeric(y5) > as.numeric(y4)) & y5 != 'NA'))
# nrow(prm_later_and_lvl3_yr5)
# 
# # not promoted later, not level 3 in 5 years
# notprm_later_and_notlvl3_yr5 <- vet.seq.10 %>%
#   filter((!(as.numeric(y5) >= 3) & y5 != 'NA') & (!((as.numeric(y10) > as.numeric(y9)) & y10 != 'NA') |((as.numeric(y9) > as.numeric(y8)) & y9 != 'NA') |((as.numeric(y7) > as.numeric(y6)) & y7 != 'NA') |(as.numeric(y5) > as.numeric(y4)) & y5 != 'NA'))
# nrow(notprm_later_and_notlvl3_yr5)
# 
# # promoted in first period ("early performers") reaching lower management in 5 years (level 3)
# # vs "late performers" doing so
# tblD <- matrix(c(129, 523, 16, 440), nrow = 2, ncol = 2)
# chisq.test(tblD)

# add job zone 5

```
